refactor: move new class to existing folder
