+++ extensions/podman/packages/extension/src/extension.ts
@@ -41,9 +41,9 @@ import { PodmanInfoHelper } from './helpers/podman-info-helper';
 import { QemuHelper } from './helpers/qemu-helper';
 import { WslHelper } from './helpers/wsl-helper';
 import { PodmanInstall } from './installer/podman-install';
-import { NotificationManager } from './notification-manager/notification-manager';
 import { PodmanRemoteConnections } from './remote/podman-remote-connections';
 import { getSocketCompatibility } from './utils/compatibility-mode';
+import { NotificationManager } from './utils/notification-manager';
 import type { InstalledPodman } from './utils/podman-cli';
 import { getPodmanCli, getPodmanInstallation } from './utils/podman-cli';
 import { PodmanConfiguration } from './utils/podman-configuration';
